+++ Examples/+BrDemo/Autotrans_tutorial.m
@@ -10,36 +10,32 @@ B.PrintAll % print available signals and parameters
 B.SetTime(0:.01:30); B.SetParam({'throttle_u0'}, 100);
 B.Sim(); B.PlotSignals({'throttle', 'RPM', 'speed', 'gear'});
 
-%% Describes complex driving scenarios 
+%% Checks a property: The speed is never below 30 while in gear3 
+STL_ReadFile('Autotrans_spec.stl');
+B.PlotRobustSat(gear3_and_speed_low)
+
+
+%% Describes and generate driving scenarios 
 % We create an input generator that will alternates between acceleration and braking 
 sg = var_step_signal_gen({'throttle', 'brake'}, 5);
 B.SetInputGen(sg);
-
-%%
+            
 % We assign ranges for duration and amplitude of each input:
 B.SetParamRanges({'dt_u0', 'dt_u1', 'dt_u2', 'dt_u3'}, ...
                   [.1 10  ;  .1 10;    0.1 10;    0.1 10]);
 B.SetParamRanges({'throttle_u0','brake_u1', 'throttle_u2', 'brake_u3'}, ... 
                   [0 100;        0 325;      0 100;         0 325]);
+
 % We don't specify a range for brake_u0 so that it remains constant equal
 % to 0 (by default). Same for throttle_u1, etc.
-              
-              
-%% Run one simulation and plots result
-% For each parameter for which we assigned a range, Breach will pick the
-% center value (e.g., throttle_u0=50, etc)
-B.Sim();
-B.PlotSignals({'throttle', 'brake','RPM', 'speed', 'gear'});
 
-%% Checks a property: The speed is never below 30 while in gear3 
-STL_ReadFile('Autotrans_spec.stl');
-B.PlotRobustSat(gear3_and_speed_low)
-
-%% Generate and run multiple test cases
 B.QuasiRandomSample(10); B.Sim();
+
+%% Plot multiple simaltions result
 B.PlotSignals({'throttle', 'brake','RPM', 'speed', 'gear'});
 
 %% Check property visually 
+figure;
 B.PlotSigPortrait({'gear', 'speed'})   
 
 %% Checks property by monitoring 
@@ -47,6 +43,7 @@ B.CheckSpec(never_gear3_and_speed_low);
 B.PrintSpecs
 
 %% Falsify property
+B.ResetSampling(); % remove the 10 samples and traces, keep parameter ranges
 falsif_pb = FalsificationProblem(B, never_gear3_and_speed_low);
 falsif_pb.max_time = 180; % give the solver three minutes to falsify the property
 falsif_pb.solve();